requirements.txt
fastapi
uvicorn
pillow
python-multipart
jinja2
tensorflow
numpy
pytest